refactor(auth): move mutation hooks to API layer, converte auth icons to SVG format, and move them to `public` folder, remove unnecessary props in `OAuthButton`.
