+++ src/features/auth/oauth-login/ui/OAuthButton.tsx
@@ -1,44 +1,48 @@
 import type { ButtonHTMLAttributes } from 'react';
-import { Button, buttonVariants } from 'shared/ui';
+import { Button } from 'shared/ui';
 import { cn } from 'shared/lib/utils';
 import { CAuth } from 'entities/auth';
 import type { TAuth } from 'entities/auth';
-import { type VariantProps } from 'class-variance-authority';
 import Link from 'next/link';
 import { type Route } from 'next';
+import Image from 'next/image';
 
-type OAuthButtonProps = Omit<ButtonHTMLAttributes<HTMLButtonElement>, 'children'> &
-  VariantProps<typeof buttonVariants> & {
-    provider: TAuth.OAuthProvider;
-    iconClassName?: string;
-    href: Route;
-  };
+type OAuthButtonProps = Omit<ButtonHTMLAttributes<HTMLButtonElement>, 'children'> & {
+  provider: TAuth.OAuthProvider;
+  iconClassName?: string;
+  href: Route;
+};
 
 export function OAuthButton({
   provider,
   className,
   iconClassName,
-  size = 'icon',
-  variant = 'outline',
   href,
   ...props
 }: OAuthButtonProps) {
   const meta = CAuth.OAUTH_PROVIDERS[provider];
-
   if (!meta) return null;
-
-  const Icon = meta.icon;
 
   return (
     <Button
       type="button"
-      size={size}
-      variant={variant}
+      size={meta.iconSrc ? 'icon' : 'default'}
+      variant={'outline'}
       className={cn(meta.buttonClassName, className)}
       {...props}
     >
-      <Link href={href}>
-        <Icon className={cn('size-6', iconClassName)} />
+      <Link href={href} className="text-base">
+        {meta.iconSrc ? (
+          <Image
+            src={meta.iconSrc}
+            alt={meta.label}
+            width={24}
+            height={24}
+            className={cn('size-6', iconClassName)}
+          />
+        ) : (
+          meta.label
+        )}
       </Link>
     </Button>
   );
