+++ src/pages/profile/api/useConnectOauthProvider.ts
@@ -1,5 +1,5 @@
 import { type DefaultError, useMutation, UseMutationOptions } from '@tanstack/react-query';
-import { AuthHttp, TAuth } from 'entities/auth';
+import { AuthHttp, type TAuth } from 'entities/auth';
 
 export type UseConnectOAuthProviderOptions = Omit<
   UseMutationOptions<TAuth.ConnectOAuthProviderResponse, DefaultError, TAuth.OAuthProvider>,
