+++ src/pages/profile/api/useDisconnectOauthProvider.ts
@@ -1,5 +1,5 @@
 import { type DefaultError, useMutation, UseMutationOptions } from '@tanstack/react-query';
-import { AuthHttp, TAuth } from 'entities/auth';
+import { AuthHttp, type TAuth } from 'entities/auth';
 
 export type UseDisconnectOAuthProviderOptions = Omit<
   UseMutationOptions<TAuth.RemoveOAuthProviderResponse, DefaultError, TAuth.OAuthProvider>,
