+++ src/pages/profile/ui/me-page/account-section/AccountsSection.tsx
@@ -1,5 +1,5 @@
 import { OAuthManageButton } from './OAuthManageButton';
-import { useConnectedAccounts } from 'pages/profile/model/useConnectedAccounts';
+import { useConnectedAccounts } from 'pages/profile/api/useConnectedAccounts';
 import { CardSection } from 'shared/ui';
 
 export function AccountSection() {
